Separation of proc macros, nightly clippy and some tiny changes
