+++ examples/wasm/Cargo.toml
@@ -2,6 +2,7 @@
 name = "wasm"
 version = "0.1.0"
 edition = "2021"
+publish = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
